Create Addons for Summernote
Create addons for summernotes.  Addons attach to the top or bottom of
your summernote editor.  We've implemented colResizable jquery code as
an add on to allow for table resizing in the summernote editor.

Made a minor "readability" change to readme.

Exposed Dom, Range, and Renderer summernote objects to enable external
development and customization.

Added 2 new predicate node tests.